Перенес получение LocaleState в SignInViewModel.kt
