--- app/src/main/java/com/example/traindriver/domain/use_case/SaveLocaleInLocalStorageUseCase.kt
@@ -1,9 +1,0 @@
-package com.example.traindriver.domain.use_case
-
-import com.example.traindriver.ui.util.LocaleState
-
-class SaveLocaleInLocalStorageUseCase {
-    suspend fun execute(localeState: LocaleState) {
-        TODO("Сохранение localeState в Room")
-    }
-}
