+++ app/src/main/java/com/example/traindriver/ui/screen/splash_screen/SplashViewModel.kt
@@ -1,6 +1,5 @@
 package com.example.traindriver.ui.screen.splash_screen
 
-import android.util.Log
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
@@ -8,18 +7,15 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.traindriver.PreferencesApp
 import com.example.traindriver.data.repository.DataStoreRepository
-import com.example.traindriver.domain.use_case.GetLocaleUseCase
-import com.example.traindriver.domain.use_case.SaveLocaleInLocalStorageUseCase
 import com.example.traindriver.ui.screen.ScreenEnum
-import com.example.traindriver.ui.util.LocaleState
-import kotlinx.coroutines.*
+import kotlinx.coroutines.coroutineScope
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.launch
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
 
 class SplashViewModel : ViewModel(), KoinComponent {
-    private val getLocaleUseCase: GetLocaleUseCase by inject()
     private val repository: DataStoreRepository by inject()
-    private val saveLocaleUseCase: SaveLocaleInLocalStorageUseCase by inject()
 
     private val _isLoading: MutableState<Boolean> = mutableStateOf(true)
     val isLoading: State<Boolean> = _isLoading
@@ -27,35 +23,20 @@ class SplashViewModel : ViewModel(), KoinComponent {
     private val _startDestination: MutableState<String> = mutableStateOf(ScreenEnum.SIGN_IN.name)
     val startDestination: State<String> = _startDestination
 
-    val locale: MutableState<LocaleState> = mutableStateOf(LocaleState.OTHER)
-
-    val number: MutableState<String> = mutableStateOf(LocaleState.OTHER.prefix())
-    val allowEntry: MutableState<Boolean> = mutableStateOf(true)
-
     init {
         viewModelScope.launch {
-            try {
-                withTimeout(PreferencesApp.startLoading) {
-                    initSetting()
-                }
-            } catch (e: TimeoutCancellationException) {
-                _isLoading.value = false
-            } catch (e: Throwable) {
-                Log.d("ZZZ", "$e")
-                _isLoading.value = false
-            }
+            initSetting()
         }
     }
 
     private suspend fun initSetting() = coroutineScope {
-        val isRegistered = async { getRegisteredState() }.run {
-            this.await()
+        launch {
+            getRegisteredState()
         }
-        if (!isRegistered) {
-            launch { getLocale() }.join()
-            launch { saveLocaleUseCase.execute(locale.value) }.join()
+        launch {
+            delay(PreferencesApp.startLoading)
+            _isLoading.value = false
         }
-        _isLoading.value = false
     }
 
 
@@ -68,14 +49,5 @@ class SplashViewModel : ViewModel(), KoinComponent {
             _startDestination.value = ScreenEnum.SIGN_IN.name
             false
         }
-    }
-
-    private suspend fun getLocale() = coroutineScope {
-        getLocaleUseCase.execute()
-            .collect { localeState ->
-                locale.value = localeState
-                number.value = localeState.prefix()
-                allowEntry.value = localeState == LocaleState.OTHER
-            }
     }
 }
